pyproject.toml
[project]
name = "audio-mixer"
version = "0.1.0"
description = "Streamlit app for mixing video audio streams with adjustable volumes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "streamlit>=1.52.2",
]
